Support 32-bit ARM on 1.5+

+++ goid_go1.5.go
@@ -16,7 +16,7 @@ type gobuf struct {
 	pc   uintptr
 	g    uintptr
 	ctxt uintptr
-	ret  uint64
+	ret  uintptr
 	lr   uintptr
 	bp   uintptr
 }
